Added details for installation both globally and as a part of a project.
Also added a nice bit about using the Symfony2 standard as part of a Symfony2
project. It's not as easy as it would seem. With the method outlined in the README,
you can remove the following line from the standard's ruleset.xml, as it doesn't
belong as part of the standard, but is rather project specific.